requirements.txt
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
asyncgui==0.6.3
asynckivy==0.6.4
attrs==23.2.0
awsebcli==3.21.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
boto3==1.35.69
botocore==1.35.69
buildozer==1.5.0
cement==2.10.14
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.3
django-cors-headers==4.6.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.16.1
fonttools==4.54.1
fqdn==1.5.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
ipykernel==6.29.4
ipython==8.22.2
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-examples==2.3.0
Kivy-Garden==0.1.5
kivy_deps.gstreamer==0.3.4
kivymd @ git+https://github.com/kivymd/KivyMD.git@5ff9d0de78260383fae0737716879781257155a8
kiwisolver==1.4.7
MarkupSafe==2.1.5
materialyoucolor==2.0.9
matplotlib==3.9.2
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.26.4
overrides==7.7.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pathspec==0.10.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pygame==2.6.1
Pygments==2.17.2
PyJWT==2.10.0
pyparsing==3.1.4
pypiwin32==223
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
s3transfer==0.10.4
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.2
setuptools==75.3.0
sh==2.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.2
stack-data==0.6.3
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.2.1
tk==0.1.0
tkcalendar==1.6.1
tornado==6.4
traitlets==5.14.2
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.20
utils==1.0.2
virtualenv==20.27.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
wheel==0.44.0
widgetsnbextension==4.0.10
